# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant clinic management: auth, tenant registry, and appointments. Except for /auth/register, /auth/login and /auth/refresh, every endpoint requires a JWT bearer token (BearerAuth). Tenant-scoped endpoints additionally require the X-Tenant-ID header (see each operation).",
        "title": "Clinic Management API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/appointments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists appointments in the active tenant with optional filters and pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "List appointments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant id",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by patient id",
                        "name": "patient_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by doctor id",
                        "name": "doctor_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "scheduled",
                            "confirmed",
                            "completed",
                            "cancelled",
                            "no_show"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start of time window (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End of time window (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Page offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates an appointment in the active tenant (X-Tenant-ID). Concurrency-safe against overlapping schedules; providing the Idempotency-Key header makes the request repeatable — a replayed request returns the stored response with an Idempotent-Replay: true header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Book an appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant id",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client-generated idempotency key",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Appointment details",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_service.BookInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/appointments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single appointment by id within the active tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Get appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant id",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Appointment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/appointments/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels an appointment in the active tenant with a required reason.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Cancel appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant id",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Appointment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cancellation reason",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CancelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/appointments/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks an appointment as completed. Requires the admin or staff role in the active tenant; patients may not perform this action.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Complete appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant id",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Appointment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/appointments/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks an appointment as confirmed. Requires the admin or staff role in the active tenant; patients may not perform this action.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Confirm appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant id",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Appointment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/appointments/{id}/no-show": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks an appointment as a no-show. Requires the admin or staff role in the active tenant; patients may not perform this action.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Mark appointment as no-show",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant id",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Appointment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/appointments/{id}/reschedule": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Moves an appointment to a new start time and/or duration in the active tenant.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Reschedule appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant id",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Appointment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New timing",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_service.RescheduleInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperr.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apperr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse": {
            "type": "object",
            "properties": {
                "cancellation_reason": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentsListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_PandaX185_clinic-management_internal_scheduling_api.appointmentResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_PandaX185_clinic-management_internal_scheduling_service.BookInput": {
            "type": "object",
            "required": [
                "doctor_id",
                "duration_minutes",
                "patient_id",
                "start_time"
            ],
            "properties": {
                "doctor_id": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer",
                    "maximum": 480,
                    "minimum": 5
                },
                "notes": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "github_com_PandaX185_clinic-management_internal_scheduling_service.RescheduleInput": {
            "type": "object",
            "required": [
                "duration_minutes",
                "start_time"
            ],
            "properties": {
                "duration_minutes": {
                    "type": "integer",
                    "maximum": 480,
                    "minimum": 5
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "service.CancelInput": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 1000
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Authenticate as `Bearer \u003caccess_token\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}